requirements.txt
streamlit==1.36.0
pymongo==4.8.0
faker==30.8.0
python-dotenv==1.0.1
